+++ Sources/Navigator/EPUB/EPUBReflowableSpreadView.swift
@@ -86,10 +86,15 @@ final class EPUBReflowableSpreadView: EPUBSpreadView {
     }
 
     private func updateContentInset() {
+        // We use the window's safeAreaInsets instead of the view's because we
+        // only want to take into account the device notch and status bar, not
+        // the application's bars.
+        let safeAreaInsets = window?.safeAreaInsets ?? UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
+        
         if viewModel.scroll {
             topConstraint.constant = 0
             bottomConstraint.constant = 0
-            scrollView.contentInset = UIEdgeInsets(top: notchAreaInsets.top, left: 0, bottom: notchAreaInsets.bottom, right: 0)
+            scrollView.contentInset = UIEdgeInsets(top: safeAreaInsets.top, left: 0, bottom: safeAreaInsets.bottom, right: 0)
 
         } else {
             let contentInset = viewModel.config.contentInset
@@ -98,9 +103,10 @@ final class EPUBReflowableSpreadView: EPUBSpreadView {
                 ?? contentInset[.unspecified]
                 ?? (top: 0, bottom: 0)
 
-            // Increases the insets by the notch area (eg. iPhone X) to make sure that the content is not overlapped by the screen notch.
-            insets.top += notchAreaInsets.top
-            insets.bottom += notchAreaInsets.bottom
+            // Increases the insets by the window's safe area insets area to
+            // make sure that the content is not overlapped by the screen notch.
+            insets.top += safeAreaInsets.top
+            insets.bottom += safeAreaInsets.bottom
 
             topConstraint.constant = insets.top
             bottomConstraint.constant = -insets.bottom