+++ Sources/Navigator/PDF/PDFDocumentView.swift
@@ -40,9 +40,13 @@ public final class PDFDocumentView: PDFView {
     }
 
     private func updateContentInset() {
-        // Setting the horizontal values triggers shifts the content incorrectly, somehow.
-        firstScrollView?.contentInset.top = notchAreaInsets.top
-        firstScrollView?.contentInset.bottom = notchAreaInsets.bottom
+        // We use the window's safeAreaInsets instead of the view's because we
+        // only want to take into account the device notch and status bar, not
+        // the application's bars.
+        let insets = window?.safeAreaInsets ?? UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
+
+        firstScrollView?.contentInset.top = insets.top
+        firstScrollView?.contentInset.bottom = insets.bottom
     }
 
     override public func canPerformAction(_ action: Selector, withSender sender: Any?) -> Bool {